requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
requests
dj-database-url
supabase
